+++ Categories/NSDate+Day.m
@@ -42,7 +42,7 @@
     return comps.minute == 0;
 }
 
-- (int)timeInMinutesSinceMidnight
+- (NSInteger)timeInMinutesSinceMidnight
 {
 	NSCalendar *gregorian = [NSCalendar currentCalendar];
 	unsigned unitflags = NSYearCalendarUnit | NSMonthCalendarUnit | NSDayCalendarUnit;
@@ -53,11 +53,11 @@
     return ([self timeIntervalSince1970] - [midnight timeIntervalSince1970]) / 60;
 }
 
-+(int)findPreviousInterval:(int) minutesFromMidnight {
++(NSInteger)findPreviousInterval:(NSInteger) minutesFromMidnight {
 	return [self findNearestInterval:minutesFromMidnight] - 15;
 }
 
-+(int)findNearestInterval:(int) minutesFromMidnight {
++(NSInteger)findNearestInterval:(NSInteger) minutesFromMidnight {
 	int numIntervals = floor(minutesFromMidnight / 15);
 	int remainder = minutesFromMidnight % 15;
 	if (remainder >= 8) {
